Refactor test class structure
