+++ tests/Microsoft.DotNet.Framework.Docker.Tests/Config.cs
@@ -1,0 +1,32 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.IO;
+using System.Text.RegularExpressions;
+using Newtonsoft.Json.Linq;
+
+namespace Microsoft.DotNet.Framework.Docker.Tests
+{
+    public static class Config
+    {
+        public static bool IsLocalRun = Environment.GetEnvironmentVariable("LOCAL_RUN") != null;
+        public static string OSFilter => Environment.GetEnvironmentVariable("IMAGE_OS_FILTER");
+        public static string RepoPrefix { get; } = Environment.GetEnvironmentVariable("REPO_PREFIX") ?? string.Empty;
+        public static string Registry { get; } = Environment.GetEnvironmentVariable("REGISTRY") ?? GetManifestRegistry();
+        public static string VersionFilter => Environment.GetEnvironmentVariable("IMAGE_VERSION_FILTER");
+
+        public static string GetManifestRegistry()
+        {
+            string manifestJson = File.ReadAllText("manifest.json");
+            JObject manifest = JObject.Parse(manifestJson);
+            return (string)manifest["registry"];
+        }
+
+        public static string GetFilterRegexPattern(string filter)
+        {
+            return $"^{Regex.Escape(filter).Replace(@"\*", ".*").Replace(@"\?", ".")}$";
+        }
+    }
+}
