+++ tests/Microsoft.DotNet.Framework.Docker.Tests/OsVersion.cs
@@ -1,0 +1,14 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Microsoft.DotNet.Framework.Docker.Tests
+{
+    public static class OsVersion
+    {
+        public const string WSC_LTSC2016 = "windowsservercore-ltsc2016";
+        public const string WSC_1803 = "windowsservercore-1803";
+        public const string WSC_LTSC2019 = "windowsservercore-ltsc2019";
+        public const string WSC_1903 = "windowsservercore-1903";
+    }
+}
